requirements.txt
# Core dependencies
streamlit==1.32.0
google-cloud-aiplatform==1.47.0
pandas==2.2.1
plotly==5.20.0

# Optional (if you add notebooks later)
jupyter==1.0.0
matplotlib==3.8.3
seaborn==0.13.2

# For data collection (if sharing scraping code)
requests==2.31.0
beautifulsoup4==4.12.3